+++ src/savingfunds/cli.py
@@ -19,6 +19,7 @@ from savingfunds.commands.edit_commands import (
     rename_account,
     rename_fund,
     set_balance,
+    change_iban,
 )
 from savingfunds.commands.money_commands import deposit, withdraw
 from savingfunds.commands.new_commands import (
@@ -97,6 +98,7 @@ cli.add_command(rename_account)
 cli.add_command(change_monthly_factor)
 cli.add_command(change_account)
 cli.add_command(change_parent_group)
+cli.add_command(change_iban)
 
 cli.add_command(remove_account)
 cli.add_command(remove_fund)